fixed the flag issue
