+++ e2etests/e2e_calls.go
@@ -4,8 +4,8 @@ import (
 	apicalls "github.com/xaaha/hulak/pkg/apiCalls"
 )
 
-func RunFormData(envMap map[string]string) {
-	apicalls.SendApiRequest(envMap, "e2etests/test_collection/form_data.yaml")
+func RunFormData(envMap map[string]string, fp string) {
+	apicalls.SendApiRequest(envMap, fp)
 }
 
 func RunFormDataError(envMap map[string]string) {
@@ -15,3 +15,26 @@ func RunFormDataError(envMap map[string]string) {
 func RunUrlEncodedFormData(envMap map[string]string) {
 	apicalls.SendApiRequest(envMap, "e2etests/test_collection/url_encoded_form.yaml")
 }
+
+/*
+func main() {
+	envMap := InitializeProject()
+	var wg sync.WaitGroup
+	// apicalls.TestApiCalls() // temp call.. replace with mock
+
+	tasks := []func(map[string]string){
+		e2etests.RunFormData,
+		e2etests.RunUrlEncodedFormData,
+		e2etests.RunFormDataError,
+	}
+
+	for _, task := range tasks {
+		wg.Add(1) // one for each task
+		go func(taskfunc func(map[string]string), env map[string]string) {
+			defer wg.Done()
+			taskfunc(utils.CopyEnvMap(env))
+		}(task, envMap)
+	}
+	wg.Wait()
+}
+*/
